reorg(1/5): move Home models to models/
Relocates `HomeDeviceModel` and `HomeDeviceRegistry` from
`src/services/` (where they lived for historical reasons) to
`src/models/`, alongside their Classic counterparts (`AreaModel`,
`BuildingModel`, `DeviceModel`, `FloorModel`). The two families of
models are conceptually identical — domain entities with identity,
mutable state, and sync semantics — and belong in the same folder.

Files:
- `services/home-device-model.ts`    → `models/home-device.ts`
- `services/home-device-registry.ts`  → `models/home-registry.ts`

The public `main.ts` re-exports `HomeDeviceModel` / `HomeDeviceRegistry`
from `./models/index.ts` now instead of `./services/index.ts`. Consumer-
visible API is unchanged (same class names, same module entry point).
