+++ src/facades/home-device-ata.ts
@@ -1,4 +1,4 @@
-import type { HomeDeviceModel } from '../services/home-device-model.ts'
+import type { HomeDeviceModel } from '../models/home-device.ts'
 import type { HomeAPI } from '../services/interfaces.ts'
 import type {
   HomeAtaValues,
