+++ src/facades/home-manager.ts
@@ -1,4 +1,4 @@
-import type { HomeDeviceModel } from '../services/home-device-model.ts'
+import type { HomeDeviceModel } from '../models/home-device.ts'
 import type { HomeAPI } from '../services/interfaces.ts'
 import { HomeDeviceAtaFacade } from './home-device-ata.ts'
