+++ src/models/home-device.ts
