+++ src/services/index.ts
@@ -23,5 +23,3 @@ export type {
 
 export { MELCloudAPI } from './api.ts'
 export { MELCloudHomeAPI } from './home-api.ts'
-export { HomeDeviceModel } from './home-device-model.ts'
-export { HomeDeviceRegistry } from './home-device-registry.ts'
