+++ tests/unit/home-device-ata-facade.test.ts
@@ -4,7 +4,7 @@ import type { HomeAPI } from '../../src/services/interfaces.ts'
 import type { HomeDevice } from '../../src/types/index.ts'
 import { HomeDeviceType } from '../../src/constants.ts'
 import { HomeDeviceAtaFacade } from '../../src/facades/home-device-ata.ts'
-import { HomeDeviceModel } from '../../src/services/home-device-model.ts'
+import { HomeDeviceModel } from '../../src/models/home-device.ts'
 import { mock } from '../helpers.ts'
 
 const defaultCapabilities: HomeDevice['capabilities'] = {
